[android] move Other preferences to dedicated file using Abstract base class
