+++ android/res/xml/settings_application.xml
@@ -53,7 +53,7 @@
 
     <android.support.v7.preference.PreferenceCategory android:title="@string/other_settings">
         <android.support.v7.preference.Preference
-            android:fragment="com.frostwire.android.gui.activities.SettingsActivity2$Other"
+            android:fragment="com.frostwire.android.gui.fragments.preference.OtherFragment"
             android:title="@string/notification_other_header" />
         <android.support.v7.preference.Preference
             android:layout="@layout/view_preference_buy_ads_2"
